implement use register, but there are some bugs.

+++ cmd/api/rpc/user.go
@@ -16,53 +16,53 @@
 package rpc
 
 import (
-  "context"
-  "github.com/Yra-A/Douyin_Simple_Demo/pkg/middleware"
-  "time"
+	"context"
+	"github.com/Yra-A/Douyin_Simple_Demo/pkg/middleware"
+	"time"
 
-  "github.com/Yra-A/Douyin_Simple_Demo/kitex_gen/user"
-  "github.com/Yra-A/Douyin_Simple_Demo/kitex_gen/user/userservice"
-  "github.com/Yra-A/Douyin_Simple_Demo/pkg/constants"
-  "github.com/Yra-A/Douyin_Simple_Demo/pkg/errno"
-  "github.com/cloudwego/kitex/client"
-  "github.com/cloudwego/kitex/pkg/retry"
-  etcd "github.com/kitex-contrib/registry-etcd"
-  trace "github.com/kitex-contrib/tracer-opentracing"
+	"github.com/Yra-A/Douyin_Simple_Demo/kitex_gen/user"
+	"github.com/Yra-A/Douyin_Simple_Demo/kitex_gen/user/userservice"
+	"github.com/Yra-A/Douyin_Simple_Demo/pkg/constants"
+	"github.com/Yra-A/Douyin_Simple_Demo/pkg/errno"
+	"github.com/cloudwego/kitex/client"
+	"github.com/cloudwego/kitex/pkg/retry"
+	etcd "github.com/kitex-contrib/registry-etcd"
+	trace "github.com/kitex-contrib/tracer-opentracing"
 )
 
 var userClient userservice.Client
 
 func initUserRpc() {
-  r, err := etcd.NewEtcdResolver([]string{constants.EtcdAddress})
-  if err != nil {
-    panic(err)
-  }
+	r, err := etcd.NewEtcdResolver([]string{constants.EtcdAddress}) // 服务发现
+	if err != nil {
+		panic(err)
+	}
 
-  c, err := userservice.NewClient(
-    constants.UserServiceName,
-    client.WithMiddleware(middleware.CommonMiddleware),
-    client.WithInstanceMW(middleware.ClientMiddleware),
-    client.WithMuxConnection(1),                       // mux
-    client.WithRPCTimeout(3*time.Second),              // rpc timeout
-    client.WithConnectTimeout(50*time.Millisecond),    // conn timeout
-    client.WithFailureRetry(retry.NewFailurePolicy()), // retry
-    client.WithSuite(trace.NewDefaultClientSuite()),   // tracer
-    client.WithResolver(r),                            // resolver
-  )
-  if err != nil {
-    panic(err)
-  }
-  userClient = c
+	c, err := userservice.NewClient(
+		constants.UserServiceName,
+		client.WithMiddleware(middleware.CommonMiddleware),
+		client.WithInstanceMW(middleware.ClientMiddleware),
+		client.WithMuxConnection(1),                       // mux
+		client.WithRPCTimeout(3*time.Second),              // rpc timeout
+		client.WithConnectTimeout(50*time.Millisecond),    // conn timeout
+		client.WithFailureRetry(retry.NewFailurePolicy()), // retry
+		client.WithSuite(trace.NewDefaultClientSuite()),   // tracer
+		client.WithResolver(r),                            // resolver
+	)
+	if err != nil {
+		panic(err)
+	}
+	userClient = c
 }
 
-// UserRegister 用户注册 rpc 客户端
+// UserRegister 用户注册【rpc 客户端】
 func UserRegister(ctx context.Context, req *user.UserRegisterRequest) error {
-  resp, err := userClient.UserRegister(ctx, req)
-  if err != nil {
-    return err
-  }
-  if resp.StatusCode != 0 {
-    return errno.NewErrNo(resp.StatusCode, *resp.StatusMsg)
-  }
-  return nil
+	resp, err := userClient.UserRegister(ctx, req)
+	if err != nil {
+		return err
+	}
+	if resp.StatusCode != 0 {
+		return errno.NewErrNo(int64(resp.StatusCode), *resp.StatusMsg)
+	}
+	return nil
 }
